+++ src/store/src/Bridge/SurrealDb/Store.php
@@ -112,6 +112,15 @@ class Store implements ManagedStoreInterface, StoreInterface
         $this->request('DELETE', \sprintf('key/%s', $this->table), []);
     }
 
+    public function count(): int
+    {
+        $this->authenticate();
+
+        $result = $this->request('POST', 'sql', \sprintf('SELECT count() FROM %s GROUP ALL;', $this->table));
+
+        return $result[0]['result'][0]['count'] ?? 0;
+    }
+
     /**
      * @param array<string, mixed>|string $payload
      *