+++ src/store/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+0.10
+----
+
+ * Add `count` method to `StoreInterface`
+
 0.9
 ---
 
@@ -13,7 +18,6 @@ CHANGELOG
 
 0.7
 ---
-
  * Add `TraceableStore` profiler decorator moved from AI Bundle
  * Add `RstLoader` and `RstToctreeLoader` for loading RST files and following toctree directives
  * Add pre-query event dispatching for query enhancement before vectorization