+++ src/store/src/Bridge/ManticoreSearch/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+0.10
+----
+
+ * Add `count` method
+
 0.1
 ---