+++ src/store/src/Bridge/Postgres/CHANGELOG.md
@@ -1,8 +1,14 @@
 CHANGELOG
 =========
 
+0.10
+----
+
+ * Add `count` method
+
 0.9
 ---
+
  * Introduce `StoreFactory`
  * Introduce a `lang` argument to configure the lang used in embeddings